flask-server/img-story/requirements.txt
transformers
torch
langchain
python-dotenv
requests
streamlit
tf-keras
